Restores assignNights functionality, but some change at sometime made it print ids and not names both in chat and in the db
